requirements.txt
fastapi==0.128.0
boto3==1.42.23
python-multipart==0.0.21
python-dotenv==1.2.1
uvicorn==0.40.0
python-dotenv